feat: production hardening — CORS, indexes, security headers, redis sessions
- Restrict CORS origins via CORS_ALLOWED_ORIGINS env variable (config/cors.php)
- Add 3 missing DB indexes: pull_request_id, applications.status, servers.ip
- Add Nginx security headers (X-Frame-Options, CSP, HSTS, etc.)
- Switch SESSION_DRIVER to redis in .env.production
- Update production readiness audit (62% → 72%)
- Include accumulated model/frontend/route changes from prior sessions

+++ app/Jobs/BackupRestoreTestJob.php
@@ -102,7 +102,7 @@ class BackupRestoreTestJob implements ShouldBeEncrypted, ShouldQueue
             };
 
             if ($success) {
-                $duration = now()->diffInSeconds($this->startTime);
+                $duration = (int) now()->diffInSeconds($this->startTime);
                 $this->execution->update([
                     'restore_test_status' => 'success',
                     'restore_test_message' => 'Restore test completed successfully',
@@ -499,7 +499,7 @@ class BackupRestoreTestJob implements ShouldBeEncrypted, ShouldQueue
 
     private function markFailed(string $message): bool
     {
-        $duration = now()->diffInSeconds($this->startTime);
+        $duration = (int) now()->diffInSeconds($this->startTime);
         $this->execution->update([
             'restore_test_status' => 'failed',
             'restore_test_message' => $message,
